% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_label}
\alias{cor_label}
\title{Get label from correlation tests}
\usage{
cor_label(h)
}
\arguments{
\item{h}{An \code{htest} object (typical class from correlation
test in native R system using \code{cor.test()}).}
}
\description{
Get the manuscript-ready label out from correlation
tests.
}
