% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhPlayers-package.R
\docType{package}
\name{jhPlayers-package}
\alias{jhPlayers}
\alias{jhPlayers-package}
\title{Data from the judo-handball players study}
\description{
R-package that provides the data used in the study on judo-handball players.
}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Cristian Núñez Espinosa \email{cristian.nunez@umag.cl} (\href{https://orcid.org/0000-0002-9896-7062}{ORCID})
}

}
\keyword{internal}
