% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_sd}
\alias{mean_sd}
\alias{colmean_sd}
\title{Mean and SD}
\usage{
mean_sd(x, k = 1)

colmean_sd(.data, vars, k = 1)
}
\arguments{
\item{x}{Numeric variable.}

\item{k}{Numeric indicating the number of decimal places.}

\item{.data}{Data from which obtain columns.}

\item{vars}{Variables from which compute mean and sd.}
}
\description{
Compute mean and standard deviation in the format
"Mean ± SD".
}
