% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{style.p}
\alias{style.p}
\title{Format p-values}
\usage{
style.p(p, k = 3)
}
\arguments{
\item{p}{Numeric vector of p-value(s).}

\item{k}{Number of decimal places. Defaults to 3.}
}
\description{
Format p-values
}
